pyproject.toml
[tool.poetry]
name = "datasources"
version = "0.1.0"
description = ""
authors = ["Alex Bildner <alex.bildner@sourcegraph.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygithub = "^1.59.1"
python-dotenv = "^1.0.0"
sentence-transformers = "^2.2.2"
fire = "^0.5.0"
streamlit = "^1.25.0"
pydantic = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

